+++ src/app/crm/contact/contact.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material';
 
 @Component({
   selector: 'app-contact',
@@ -7,9 +8,61 @@ import { Component, OnInit } from '@angular/core';
 })
 export class ContactComponent implements OnInit {
 
-  constructor() { }
+  files = {
+    datas: [
+      {
+        icon: 'folder',
+        name: 'main',
+        type: 'js',
+        provide: 'stbui',
+        size: '10k',
+        lastModifiedDate: '20171014'
+      },
+      {
+        icon: 'folder',
+        name: 'vendor',
+        type: 'js',
+        provide: 'stbui',
+        size: '10k',
+        lastModifiedDate: '20171014'
+      },
+      {
+        icon: 'folder',
+        name: 'polyfills',
+        type: 'js',
+        provide: 'stbui',
+        size: '10k',
+        lastModifiedDate: '20171014'
+      }
+    ]
+  };
+
+  constructor(private dialog: MatDialog) {
+  }
 
   ngOnInit() {
   }
 
+  onUpdateContact(contact) {
+    console.log(contact);
+    const dialogRef = this.dialog.open(ContactUpdateComponent, {
+      width: '400px',
+      panelClass: 'contact-dialog-form',
+      data: {name: ''}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      console.log('dialog');
+    });
+  }
 }
+
+@Component({
+  selector: 'stb-contact-upate',
+  templateUrl: './update.component.html',
+  styleUrls: ['./update.component.scss']
+})
+export class ContactUpdateComponent {
+
+}
+