+++ src/app/pages/pages.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { SharedModule } from '../shared/shared.module';
 
-import { PagesRoutingModule } from './pages.routing.';
+import { PagesRoutingModule } from './pages.routing';
 import { PagesComponent } from './pages.component';
 import { AboutComponent } from './about/about.component';
 import { ContactComponent } from './contact/contact.component';
@@ -9,9 +9,12 @@ import { ServicesComponent } from './services/services.component';
 import { BlogComponent } from './blog/blog.component';
 import { ProfileComponent } from './profile/profile.component';
 import { CollectionComponent } from './collection/collection.component';
-import { UserComponent } from './user/user.component';
 
+import { UserComponent } from './user/user.component';
 import { UserService } from './user/user.service';
+
+import { FileManagerComponent } from './file-manager/file-manager.component';
+import { ProjectComponent } from './project/project.component';
 
 @NgModule({
   imports: [
@@ -26,10 +29,13 @@ import { UserService } from './user/user.service';
     CollectionComponent,
     BlogComponent,
     ProfileComponent,
-    UserComponent
+    UserComponent,
+    FileManagerComponent,
+    ProjectComponent
   ],
   providers: [
     {provide: 'UserService', useClass: UserService},
   ]
 })
-export class PagesModule { }
+export class PagesModule {
+}